+++ lib/whois/record/parser/whois.nic.dz.rb
@@ -28,14 +28,10 @@ module Whois
       class WhoisNicDz < Base
 
         property_supported :status do
-          if content_for_scanner =~ /ETAT:\.+(.+?)\n/
-            case $1.downcase
-              when "actif" then :registered
-              else
-                Whois.bug!(ParserError, "Unknown status `#{$1}'.")
-            end
-          else
+          if available?
             :available
+          else
+            :registered
           end
         end